Fix current and selected account flow
