+++ src/locales/en/common.json
@@ -402,6 +402,7 @@
   "Signed message": "Signed message",
   "Since you want to reclaim your LSK on the new blockchain, you need to pay the fee from your new account.": "Since you want to reclaim your LSK on the new blockchain, you need to pay the fee from your new account.",
   "Slot status": "Slot status",
+  "Something went wrong. Please try again": "Something went wrong. Please try again",
   "Sorry, we couldn’t find the page you were looking for. We suggest that you return to the main dashboard.": "Sorry, we couldn’t find the page you were looking for. We suggest that you return to the main dashboard.",
   "Standby delegates": "Standby delegates",
   "Status": "Status",
@@ -461,7 +462,7 @@
   "Type": "Type",
   "Unable to connect to Lisk Service, please check the address and try again": "Unable to connect to Lisk Service, please check the address and try again",
   "Unable to connect to the node, no response from the server.": "Unable to connect to the node, no response from the server.",
-  "Unable to decrypt account, please check the password": "Unable to decrypt account, please check the password",
+  "Unable to decrypt account. Please check your password": "Unable to decrypt account. Please check your password",
   "Undo": "Undo",
   "Unlock": "Unlock",
   "Unlock LSK summary": "Unlock LSK summary",
