+++ src/modules/account/components/RemoveSuccess/RemoveSuccess.js
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 
 import { PrimaryButton } from 'src/theme/buttons';
 import Icon from 'src/theme/Icon';
-import styles from '../RemoveCurrentAccountFlow/RemoveAccount.css';
+import styles from '../RemoveAccount/RemoveAccount.css';
 
 const RemoveSuccess = ({ onComplete }) => {
   const { t } = useTranslation();
